+++ tests/testthat/test-hex.R
@@ -1,7 +1,11 @@
 test_that("read_hex() and write_hex()", {
     font_file <- system.file("fonts/spleen/spleen-8x16.hex.gz", package = "bittermelon")
     font <- read_hex(font_file)
+    expect_length(font, 837L)
     expect_true(is_bm_font(font))
+
+    font <- read_hex(font_file, ucp = block2ucp("Basic Latin"))
+    expect_length(font, 95L)
 
     expect_true(any(grepl(" * Spleen is released under the BSD 2-Clause",
                           attr(font, "comments"))))